Phase 10: release baseline, warning policy, Windows and sanitiser scaffolding
Clean from-scratch Release and Debug builds, both at zero warnings from FOUR
COLOR's own sources with -Werror on.

- CMake: warning flags attached to our translation units only, since JUCE's
  module sources compile into the same targets. -Wconversion/-Wdouble-promotion
  are deliberately excluded (262 hits, all intentional float/int arithmetic);
  the enabled set reaches a genuine zero.
- Fixed every warning it found: a shadowed member in TptSvf::setCoefficients,
  FourColorEditor's 'processor' shadowing AudioProcessorEditor's, and 15
  implicit int->float conversions now cast explicitly.
- Performance checks are Release-only. A CPU figure from an unoptimised Debug
  build is meaningless, and lowering the Release thresholds to accommodate it
  would be tuning a test to pass.
- FourColorTestsSan target: ASan + UBSan, -fno-sanitize-recover=all. Run clean,
  304 checks, no findings.
- scripts/build-macos.sh and scripts/build-windows.bat run the same steps on
  both platforms; neither touches the plug-in folders without --install.
- .github/workflows/windows.yml pins windows-2022 and the exact JUCE commit.
- docs/BUILDING.md pins the toolchain and states plainly that pluginval is not
  installed. docs/CUBASE-CHECKLIST.md is 11 sections of host validation, marked
  WINDOWS/CUBASE NOT VERIFIED throughout.

+++ Source/Dsp/ColorEngine.cpp
@@ -34,7 +34,7 @@ namespace fourcolor
 
         for (int i = 0; i < 16; ++i)
         {
-            const float phase = (i + 0.5f) * (juce::MathConstants<float>::pi / 16.0f);
+            const float phase = ((float) i + 0.5f) * (juce::MathConstants<float>::pi / 16.0f);
             const float x = ref * std::sin (phase);
             const float y = staticShape (x * preGain);
             inPower  += x * x;
